Add entrance connections to spoiler log
Also: move levellist setup to Rules to reduce unneeded imports, but keep init simplified.
